Allow @ character in argument sanitizer
It is a common practice to use SHA256 signature of container
image names, like `registry.example.com/app/web@sha256:<sig>`.
The argument sanitizer doesn't allow the `@` character, so these
image names will be truncated and will fail to be pulled.

This change adds the `@` character in the accepted characters.

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>